add unit test for coffee service

+++ src/main/java/com/jstjnsn/coffeeapi/service/CoffeeService.java
@@ -1,10 +1,10 @@
 package com.jstjnsn.coffeeapi.service;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
+import com.jstjnsn.coffeeapi.exception.CoffeeNotFoundException;
 import com.jstjnsn.coffeeapi.model.Coffee;
 import com.jstjnsn.coffeeapi.repository.CoffeeRepository;
 
@@ -20,22 +20,25 @@ public class CoffeeService {
         return repository.findAll();
     }
 
-    public Optional<Coffee> getCoffeeById(Long id) {
-        return repository.findById(id);
+    public Coffee getCoffeeById(Long id) {
+        return repository.findById(id).orElseThrow(() -> new CoffeeNotFoundException(id));
     }
 
     public Coffee createCoffee(Coffee coffee) {
         return repository.save(coffee);
     }
 
-    public Optional<Coffee> updateCoffee(Long id, Coffee updatedCoffee) {
+    public Coffee updateCoffee(Long id, Coffee updatedCoffee) {
         return repository.findById(id).map(coffee -> {
             coffee.setName(updatedCoffee.getName());
             return repository.save(coffee);
-        });
+        }).orElseThrow(() -> new CoffeeNotFoundException(id));
     }
 
     public void deleteCoffee(Long id) {
+        if (!repository.existsById(id)) {
+            throw new CoffeeNotFoundException(id);
+        }
         repository.deleteById(id);
     }
 }
